pyproject.toml
[project]
name = "jira_to_obsidian"
version = "0.1.0"
description = "Sync JIRA tickets to Obsidian vault"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "jira>=3.8.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "python-dateutil>=2.8.2",
    "click>=8.1.7",
    "rich>=13.7.0",
]

[project.scripts]
jira-to-obsidian = "jira_to_obsidian.cli:main"
j2o = "jira_to_obsidian.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/jira_to_obsidian"]

[tool.hatch.envs.default]
dependencies = [
    "black>=24.3.0",
    "ruff>=0.3.0",
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "responses>=0.24.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --cov=jira_to_obsidian --cov-report=term-missing"

[tool.coverage.run]
source = ["src/jira_to_obsidian"]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.ruff]
line-length = 88
target-version = "py311"
select = ["E", "F", "W", "N", "I"]
src = ["src", "tests"]